refactor: extract compressed backend config (#87)
