+++ crates/mcp-compressor-core/src/server/mod.rs
@@ -1,9 +1,11 @@
+pub mod backend;
 pub mod compressed;
 pub mod registration;
 pub mod tool_cache;
 
+pub use backend::{BackendAuthMode, BackendServerConfig, BackendTransport};
 pub use compressed::{
-    BackendConfigSource, BackendServerConfig, CompressedServer, CompressedServerConfig,
-    JustBashCommandSpec, JustBashProviderSpec, ProxyTransformMode,
+    BackendConfigSource, CompressedServer, CompressedServerConfig, JustBashCommandSpec,
+    JustBashProviderSpec, ProxyTransformMode,
 };
 pub use tool_cache::ToolCache;
